[diag] Go back to sleep after transmission
When "diag radio sleep" is followed by "diag send" or "diag
repeat", the radio switches to and stays in the receive mode
after a transmission. Normally, the MAC layer is responsible
for putting the radio back to sleep but the MAC layer is
bypassed when using the diag commands.

Make sure that the radio goes back to sleep after
a transmission when the sleep mode is on. This helps command
like "diag repeat" work more reliably as the transmission
is less likely to be interferred with the accidental frame
reception.

Signed-off-by: Damian Krolik <damian.krolik@nordicsemi.no>