[mdns] introduce standalone OT mDNS library
The OpenThread core provides a native mDNS implementation primarily
intended for use within the Thread stack and Border Router functions,
such as SRP Advertising or Discovery Proxy modules. There is interest
in using this mDNS implementation in other projects without including
the full Thread stack. This commit introduces a standalone OpenThread
mDNS library, `libopenthread-mdns.a`, which includes the mDNS and
related modules for integration into other projects.

`libopenthread-mdns.a` is built alongside other high-level library
outputs, such as the FTD Thread stack (`libopenthread-ftd.a`), MTD
(`libopenthread-mtd.a`), and RCP (`libopenthread-radio.a`), and is
supported in both CMake and `BUILD.gn` build files.

This commit provides a new example platform implementation under
`/examples/platforms/posix_mdns` for POSIX systems using the
standalone mDNS library. This example builds an executable,
`ot-cli-mdns`, which runs as a fully functioning mDNS entity on a
computer and provides OpenThread CLI mDNS-related commands.

A detailed `README.md` guide for the new library is also added which
covers:
- Related OT public APIs for the mDNS library,
- Platform requirements (required platform APIs) and integration
  instructions,
- Relevant `OPENTHREAD_CONFIG` options for building the standalone
  mDNS library, including descriptions and recommended values. An
  example core-config header,
  `examples/config/ot-core-config-mdns.h`, is provided.
- Build script instructions for the new library and `ot-cli-mdns`
  executable.