[routing-manager] track and log favored OMR prefix changes
This commit enhances the `OmrPrefixManager` to track and log changes
to the favored OMR prefix. The new log includes the old favored
prefix (if any) along with the new one, and provides additional
information such as its preference, whether it is a domain prefix, or
whether it matches the local OMR prefix. This replaces and enhances
the previous logs.

This new mechanism can also be used to signal (to other modules) when
the favored OMR prefix changes.