[spinel] allow RCP reset if not multipan architecture
This VerifyOrExit will prevent the RCP recovery process from resetting
the RCP. The comment highlights that given a multipan architecture, it
is not viable to reset the RCP once it is declared ready. In
non-multipan architectures, it is still viable and desirable to reset
the RCP as part of recovering it.

See https://github.com/orgs/openthread/discussions/11273