[mle] clamp child timeout to min/max range
This commit updates `SetTimeout()` to ensure the child timeout remains
within the minimum and maximum allowed values (per the specification,
the maximum value of 8 hours is used).

These limits are also enforced when the parent requests a different
timeout value in an MLE Child Update Response to the child's
request.