[posix] make `gpio-reset-device` as optional parameter in SPI radio URL
The target of this commit is to make process exit with a new exit code
when a hardware reset is needed to be performed by POSIX platform
instead of openthread itself. For example, for some Wifi/BT/Thread
trinity chip, it can be platform or Wifi/BT stack responsibility to
hardware reset the trinity chip.

In this commit:
* A new exit code `OT_EXIT_RCP_RESET_REQUIRED` is added;
* `gpio-reset-device` and `gpio-reset-line` can be optional parameter;
* process dies if `gpio-reset-device` is not given when TriggerReset()

Change-Id: I2ff325bb4eb608b26dabbc8e56cf16bb1243f203