[channel monitor] make scan duration configurable
This allows the scan duration of the Channel Monitor to be configurable
in order to support radios/RCPs that do not provide the channel-specific
RSSI when a zero-duration energy scan is conducted.

An example is the Espressif ESP32H2 radio co-processor, which provides
the last received RSSI regardless of channel, despite the energy scan
logic of setting the receive channel, then asking for the last RSSI,
that is executed when a zero-duration scan is requested from
Mac::EnergyScan. The Channel Monitor in the case of this RCP will report
the same last RSSI received from a Thread transmitter for every channel,
which breaks the utility of the Channel Monitor.