[csl] correct the CSL Timer time
The current CSL receiver module uses the radio time to set the `CslTimer` time.
The radio platform API and the alarm platform API do not specify that the radio
time and micro timer must use the same time source. So setting the `CslTimer`
time to the radio time may cause problems.

This CL records the difference between the radio time and the `CslTimer` time,
and then uses the radio time and this difference to set the `CslTimer` time.