[csl] use correct time basis for CSL sample time
This commit updates how the CSL sample time is tracked in `SubMac`,
ensuring that it is tracked using both radio time based on
`otPlatRadioGetNow()` and the local time (`TimeMicro::GetNow()`).
These two time sources may not necessarily be the same.

This change addresses the issue where the same time value was used for
both scheduling `mCslTimer` (which should be based on local time) and
for calls to the radio APIs (which should use radio time).