[routing-manager] process RDNSS options and track addresses (#11328)
This commit introduces a mechanism in `RoutingManager` to parse and
process Recursive DNS Server (RDNSS) options within received Router
Advertisement (RA) messages. The list of discovered RDNSS addresses
is tracked (per router) by `RoutingManager`, applying the advertised
lifetime to properly age the discovered entries.

This commit adds a new public API to retrieve the list of tracked
RDNSS addresses, along with a callback mechanism where OpenThread
users can register to be notified of changes to the RDNSS list. This
callback is invoked when any of the following changes occur:

- A new RDNSS address is advertised by a router.
- A previously discovered address is removed due to the router
  advertising it with a zero lifetime.
- A previously discovered address has aged out (its lifetime expired
  without being re-advertised).
- `RoutingManager` determines that the router advertising the address
  is now unreachable, resulting in the removal of all its associated
  entries.

This commit also adds corresponding CLI commands for the new API.

Furthermore, `test_routing_manager` is updated to include a test case
covering all the newly added behaviors in detail.