[dns-client] add support for arbitrary DNS record type query
This commit introduces a new feature in `Dns::Client` to support
sending DNS queries for arbitrary record types. Callers are notified
of received response via a callback. New methods are provided to
parse and read all records in the response. Public APIs and related
CLI commands for this new feature are also added.

The `OPENTHREAD_CONFIG_DNS_CLIENT_ARBITRARY_RECORD_QUERY_ENABLE` build
configuration is added to control this feature. This allows projects
that do not require arbitrary DNS query functionality to disable it,
avoiding its associated firmware code size overhead.

Importantly, if a retrieved record type is PTR, CNAME, or SRV, the
record data in the received response contains a DNS name which may
use DNS name compression. For these specific record types, the record
data is first decompressed such that it contains the uncompressed
DNS name. For all other record types, the record data is read and
provided as it appears in the received response message.