[secure-transport] update log mapping
The log handler currently uses LogAt which causes link error when
OpenThread logs are disabled. This commit updates the log handler to
static mapping mbedtls log level to a corresponding OpenThread log
level.

Currently mbedtls has 6 log levels: 0-5, and 5 is the most logs. This
commit updates the mapping so that the level having most logs are
aligned.