[dnssd-server] support SOA query
This commit updates the DNS-SD server to support responding to SOA
record queries for the default service domain. The recommended values
for SOA record data from RFC-8766 section 6.1 are used

A new test case is added in the `test_dns_client` unit test to cover
this newly added behavior. This test verifies multiple scenarios, in
particular, when a query is made for a name not hosted by the server.
In such cases, the response would contain no records in the answer
section but include the SOA record for the zone in the additional
section.