[dnssd-server] support SOA and NS query
This commit updates the DNS-SD server to support responding to SOA and
NS record queries (including `ANY` record type queries) for the
default service domain. The recommended values for SOA record data
from RFC-8766 section 6.1 are used. The server name included in
SOA/NS answers is derived from the Extended Address of the device,
ensuring it remains fixed and consistent over reboots as long as the
device's Extended Address stays the same.

A new detailed test case is added in the `test_dns_client` unit test
to cover all the newly added behavior. This test verifies multiple
scenarios, in particular, when a query is made for a name not hosted
by the server. In such cases, the response will contain no records in
the answer section but include the SOA or NS record for the zone in
the additional data section.