[border-agent] shorten method and variable names (#11456)
This commit updates `BorderAgent` method and variable names to use
shorter forms where possible. Specifically, the term `MeshCoP` is
removed from many variable and method names, as the `BorderAgent`
class itself is already defined within the `MeshCoP` namespace.

This commit is purely a style and naming change and contains no
modification to the code logic.