[border-agent] simplify State Bitmap preparation
This commit simplifies the preparation of the State Bitmap, which is
included in the Border Agent's TXT data using the `sb` key.

Specifically, constants for field values are now directly used to
construct the `uint32_t` bitmap, making the process more
straightforward.

Additionally, relevant constants are converted to use `static
constexpr`, avoiding the use of unnamed `enum` definitions.