[dns] Update comments DNS-types to match IANA
This adds comments to DNS types, to ensure that official names from the IANA DNS parameters registry can be found in the OT source code when starting to search from that starting point.
This will make development and code analysis easier for those not familiar with OT, or coming from the DNS world.
The IANA names (which may deviate from RFC names) are from: https://www.iana.org/assignments/dns-parameters